+++ scripts/ci/pre_commit/pre_commit_check_provider_yaml_files.py
@@ -338,6 +338,7 @@ def check_unique_provider_name(yaml_files: Dict[str, Dict]):
 def check_providers_are_mentioned_in_issue_template(yaml_files: Dict[str, Dict]):
     prefix_len = len("apache-airflow-providers-")
     short_provider_names = [d['package-name'][prefix_len:] for d in yaml_files.values()]
+    short_provider_names.remove('jira')  # exclude deprecated provider that shouldn't be in issue template
     jsonpath_expr = parse('$.body[?(@.attributes.label == "Apache Airflow Provider(s)")]..options[*]')
     with PROVIDER_ISSUE_TEMPLATE_PATH.open() as issue_file:
         issue_template = yaml.safe_load(issue_file)