+++ airflow/providers/jira/CHANGELOG.rst
@@ -24,6 +24,14 @@
 Changelog
 ---------
 
+3.1.0
+.....
+
+.. warning:: This is the last release for jira provider. The provider is discontinued.
+   Please switch to atlassian.jira provider
+
+* ``Deprecate jira provider in favor of atlassian.jira provider  (#25930)``
+
 3.0.1
 .....